feat(slack): configurable typing stages, gate reorder, e2e test split
- Add TypingStageConfig to YAML config and wire through to adapter
- Persist teamID from authTest response for downstream use
- Reorder gate check before thread ownership (fixes botMentioned bug)
- Add debug logging for bot message filtering
- Fix triplet bold regex to satisfy gocritic regexpSimplify
- Split monolithic e2e/client_e2e_test.go into helper/protocol/session files
