+++ configs/config-dev.yaml
@@ -35,7 +35,7 @@ worker:
 
 messaging:
   slack:
-    enabled: false
+    enabled: true
     bot_token: ""
     app_token: ""
     socket_mode: true
