--- e2e/client_e2e_test.go
@@ -1,939 +1,0 @@
-// Package e2e provides end-to-end tests for the HotPlex Worker Gateway
-// using the Go client SDK. It spins up a full gateway stack with simulated
-// workers to validate the complete AEP v1 protocol flow for all worker types.
-package e2e_test
-
-import (
-	"context"
-	"crypto/ecdsa"
-	"crypto/elliptic"
-	"crypto/rand"
-	"fmt"
-	"io"
-	"log/slog"
-	"net/http"
-	"net/http/httptest"
-	"strings"
-	"sync"
-	"testing"
-	"time"
-
-	"github.com/golang-jwt/jwt/v5"
-	"github.com/google/uuid"
-	"github.com/stretchr/testify/mock"
-	"github.com/stretchr/testify/require"
-
-	client "github.com/hotplex/hotplex-go-client"
-
-	"github.com/hotplex/hotplex-worker/internal/config"
-	"github.com/hotplex/hotplex-worker/internal/gateway"
-	"github.com/hotplex/hotplex-worker/internal/security"
-	"github.com/hotplex/hotplex-worker/internal/session"
-	"github.com/hotplex/hotplex-worker/internal/worker"
-	"github.com/hotplex/hotplex-worker/pkg/aep"
-	"github.com/hotplex/hotplex-worker/pkg/events"
-)
-
-// ─── Simulated Worker ──────────────────────────────────────────────────────
-
-// simulatedConn implements worker.SessionConn for test workers.
-// It produces realistic AEP events in response to input.
-type simulatedConn struct {
-	sessionID string
-	userID    string
-	recvCh    chan *events.Envelope
-	closed    bool
-}
-
-func newSimulatedConn(sessionID, userID string) *simulatedConn {
-	return &simulatedConn{
-		sessionID: sessionID,
-		userID:    userID,
-		recvCh:    make(chan *events.Envelope, 64),
-	}
-}
-
-func (c *simulatedConn) Send(_ context.Context, _ *events.Envelope) error {
-	return nil
-}
-
-func (c *simulatedConn) Recv() <-chan *events.Envelope {
-	return c.recvCh
-}
-
-func (c *simulatedConn) Close() error {
-	if c.closed {
-		return nil
-	}
-	c.closed = true
-	close(c.recvCh)
-	return nil
-}
-
-func (c *simulatedConn) UserID() string    { return c.userID }
-func (c *simulatedConn) SessionID() string { return c.sessionID }
-
-// emitEvents sends a realistic event sequence (message.start → delta → end → done)
-// through the recvCh, simulating what a real worker would produce.
-func (c *simulatedConn) emitEvents(content string) {
-	msgID := "msg_" + uuid.NewString()
-
-	c.recvCh <- events.NewEnvelope(aep.NewID(), c.sessionID, 0, events.MessageStart, events.MessageStartData{
-		ID:          msgID,
-		Role:        "assistant",
-		ContentType: "text",
-	})
-
-	// Split content into deltas for realism.
-	chunks := splitContent(content, 20)
-	for _, chunk := range chunks {
-		c.recvCh <- events.NewEnvelope(aep.NewID(), c.sessionID, 0, events.MessageDelta, events.MessageDeltaData{
-			MessageID: msgID,
-			Content:   chunk,
-		})
-	}
-
-	c.recvCh <- events.NewEnvelope(aep.NewID(), c.sessionID, 0, events.MessageEnd, events.MessageEndData{
-		MessageID: msgID,
-	})
-
-	c.recvCh <- events.NewEnvelope(aep.NewID(), c.sessionID, 0, events.Done, events.DoneData{
-		Success: true,
-		Stats:   map[string]any{"input_tokens": 10, "output_tokens": len(content) / 4},
-	})
-}
-
-func splitContent(s string, chunkSize int) []string {
-	if len(s) <= chunkSize {
-		return []string{s}
-	}
-	var chunks []string
-	for i := 0; i < len(s); i += chunkSize {
-		end := i + chunkSize
-		if end > len(s) {
-			end = len(s)
-		}
-		chunks = append(chunks, s[i:end])
-	}
-	return chunks
-}
-
-// simulatedWorker implements worker.Worker for E2E tests.
-// It accepts input and emits realistic AEP events via a simulatedConn.
-type simulatedWorker struct {
-	mu         sync.Mutex
-	workerType worker.WorkerType
-	conn       *simulatedConn
-	started    bool
-	killed     bool
-}
-
-var _ worker.Worker = (*simulatedWorker)(nil)
-
-func newSimulatedWorker(wt worker.WorkerType) *simulatedWorker {
-	return &simulatedWorker{workerType: wt}
-}
-
-func (w *simulatedWorker) Type() worker.WorkerType { return w.workerType }
-func (w *simulatedWorker) SupportsResume() bool    { return true }
-func (w *simulatedWorker) SupportsStreaming() bool { return true }
-func (w *simulatedWorker) SupportsTools() bool     { return true }
-func (w *simulatedWorker) EnvWhitelist() []string  { return nil }
-func (w *simulatedWorker) SessionStoreDir() string { return "" }
-func (w *simulatedWorker) MaxTurns() int           { return 0 }
-func (w *simulatedWorker) Modalities() []string    { return []string{"text"} }
-
-func (w *simulatedWorker) Start(_ context.Context, info worker.SessionInfo) error {
-	w.mu.Lock()
-	defer w.mu.Unlock()
-	w.conn = newSimulatedConn(info.SessionID, info.UserID)
-	w.started = true
-	return nil
-}
-
-func (w *simulatedWorker) Input(_ context.Context, content string, _ map[string]any) error {
-	w.mu.Lock()
-	conn := w.conn
-	w.mu.Unlock()
-	if conn == nil {
-		return fmt.Errorf("worker not started")
-	}
-	go conn.emitEvents(content)
-	return nil
-}
-
-func (w *simulatedWorker) Resume(_ context.Context, info worker.SessionInfo) error {
-	w.mu.Lock()
-	defer w.mu.Unlock()
-	w.conn = newSimulatedConn(info.SessionID, info.UserID)
-	w.started = true
-	return nil
-}
-
-func (w *simulatedWorker) Terminate(_ context.Context) error {
-	w.mu.Lock()
-	defer w.mu.Unlock()
-	if w.conn != nil {
-		_ = w.conn.Close()
-	}
-	return nil
-}
-
-func (w *simulatedWorker) Kill() error {
-	w.mu.Lock()
-	defer w.mu.Unlock()
-	w.killed = true
-	if w.conn != nil {
-		_ = w.conn.Close()
-	}
-	return nil
-}
-
-func (w *simulatedWorker) Wait() (int, error) {
-	return 0, io.EOF
-}
-
-func (w *simulatedWorker) Conn() worker.SessionConn {
-	w.mu.Lock()
-	defer w.mu.Unlock()
-	return w.conn
-}
-
-func (w *simulatedWorker) Health() worker.WorkerHealth {
-	return worker.WorkerHealth{
-		Type:    w.workerType,
-		Healthy: true,
-		Running: true,
-		Uptime:  "1s",
-	}
-}
-
-func (w *simulatedWorker) LastIO() time.Time {
-	return time.Now()
-}
-
-func (w *simulatedWorker) ResetContext(_ context.Context) error {
-	return nil
-}
-
-// testWorkerFactory creates simulated workers for tests.
-type testWorkerFactory struct{}
-
-func (testWorkerFactory) NewWorker(t worker.WorkerType) (worker.Worker, error) {
-	return newSimulatedWorker(t), nil
-}
-
-// ─── Mock Store ─────────────────────────────────────────────────────────────
-
-// mockStore implements session.Store for E2E tests.
-// It stores sessions in-memory, matching the testify/mock pattern from
-// internal/session/manager_test.go.
-type mockStore struct {
-	mock.Mock
-}
-
-func (m *mockStore) Upsert(ctx context.Context, info *session.SessionInfo) error {
-	args := m.Called(ctx, info)
-	return args.Error(0)
-}
-
-func (m *mockStore) Get(ctx context.Context, id string) (*session.SessionInfo, error) {
-	args := m.Called(ctx, id)
-	if args.Get(0) == nil {
-		return nil, args.Error(1)
-	}
-	return args.Get(0).(*session.SessionInfo), args.Error(1)
-}
-
-func (m *mockStore) List(ctx context.Context, limit, offset int) ([]*session.SessionInfo, error) {
-	args := m.Called(ctx, limit, offset)
-	if args.Get(0) == nil {
-		return nil, args.Error(1)
-	}
-	return args.Get(0).([]*session.SessionInfo), args.Error(1)
-}
-
-func (m *mockStore) GetExpiredMaxLifetime(ctx context.Context, now time.Time) ([]string, error) {
-	args := m.Called(ctx, now)
-	if args.Get(0) == nil {
-		return nil, args.Error(1)
-	}
-	return args.Get(0).([]string), args.Error(1)
-}
-
-func (m *mockStore) GetExpiredIdle(ctx context.Context, now time.Time) ([]string, error) {
-	args := m.Called(ctx, now)
-	if args.Get(0) == nil {
-		return nil, args.Error(1)
-	}
-	return args.Get(0).([]string), args.Error(1)
-}
-
-func (m *mockStore) DeleteTerminated(ctx context.Context, cutoff time.Time) error {
-	args := m.Called(ctx, cutoff)
-	return args.Error(0)
-}
-
-func (m *mockStore) Close() error {
-	args := m.Called()
-	return args.Error(0)
-}
-
-// ─── Test Gateway Setup ─────────────────────────────────────────────────────
-
-// testGateway holds the components of a test gateway server.
-type testGateway struct {
-	server   *httptest.Server
-	hub      *gateway.Hub
-	sm       *session.Manager
-	bridge   *gateway.Bridge
-	cfg      *config.Config
-	store    *mockStore
-	log      *slog.Logger
-	cancel   context.CancelFunc
-	jwtKey   *ecdsa.PrivateKey
-	jwtValid *security.JWTValidator
-}
-
-// setupTestGateway creates a fully wired gateway server for testing.
-// It uses simulated workers and an in-memory mock store.
-func setupTestGateway(t *testing.T) *testGateway {
-	t.Helper()
-
-	ctx, cancel := context.WithCancel(context.Background())
-
-	log := slog.Default()
-
-	cfg := config.Default()
-	cfg.Security.APIKeys = nil // dev mode: allow all
-	cfg.Security.AllowedOrigins = []string{"*"}
-	cfg.Gateway.BroadcastQueueSize = 64
-	cfg.Worker.DefaultWorkDir = "/tmp"
-	cfg.Pool.MaxSize = 20
-	cfg.Pool.MaxIdlePerUser = 10
-	cfg.Pool.MaxMemoryPerUser = 0
-
-	// Generate ES256 key for JWT testing.
-	jwtKey, _ := ecdsa.GenerateKey(elliptic.P256(), rand.Reader)
-	jwtValidator := security.NewJWTValidator(jwtKey, "")
-
-	store := new(mockStore)
-	store.Test(t)
-
-	// Allow any Upsert (session creation).
-	store.On("Upsert", mock.Anything, mock.AnythingOfType("*session.SessionInfo")).Return(nil)
-	store.On("Close").Return(nil)
-	store.On("GetExpiredMaxLifetime", mock.Anything, mock.AnythingOfType("time.Time")).Return([]string{}, nil)
-	store.On("GetExpiredIdle", mock.Anything, mock.AnythingOfType("time.Time")).Return([]string{}, nil)
-	store.On("DeleteTerminated", mock.Anything, mock.AnythingOfType("time.Time")).Return(nil)
-	store.On("List", mock.Anything, mock.AnythingOfType("int"), mock.AnythingOfType("int")).Return([]*session.SessionInfo{}, nil)
-	// Get falls back to store when session is not in Manager's in-memory map.
-	// Return not-found for all store lookups (Manager holds sessions in memory after Create).
-	store.On("Get", mock.Anything, mock.AnythingOfType("string")).Return(nil, session.ErrSessionNotFound)
-
-	sm, err := session.NewManager(ctx, log, cfg, store, nil)
-	require.NoError(t, err)
-
-	hub := gateway.NewHub(log, cfg)
-
-	sm.StateNotifier = func(ctx context.Context, sessionID string, state events.SessionState, message string) {
-		env := events.NewEnvelope(aep.NewID(), sessionID, hub.NextSeq(sessionID), events.State, events.StateData{
-			State:   state,
-			Message: message,
-		})
-		_ = hub.SendToSession(ctx, env)
-	}
-
-	handler := gateway.NewHandler(log, cfg, hub, sm, jwtValidator)
-	bridge := gateway.NewBridge(log, hub, sm, nil)
-	bridge.SetWorkerFactory(testWorkerFactory{})
-
-	auth := security.NewAuthenticator(&cfg.Security, jwtValidator)
-
-	go hub.Run()
-
-	mux := http.NewServeMux()
-	mux.Handle("/ws", hub.HandleHTTP(auth, sm, handler, bridge))
-
-	server := httptest.NewServer(mux)
-
-	tg := &testGateway{
-		server:   server,
-		hub:      hub,
-		sm:       sm,
-		bridge:   bridge,
-		cfg:      cfg,
-		store:    store,
-		log:      log,
-		cancel:   cancel,
-		jwtKey:   jwtKey,
-		jwtValid: jwtValidator,
-	}
-
-	t.Cleanup(func() {
-		cancel()
-		_ = hub.Shutdown(context.Background())
-		_ = sm.Close()
-		server.Close()
-	})
-
-	return tg
-}
-
-// wsURL returns the WebSocket URL for the test server.
-func (tg *testGateway) wsURL() string {
-	return "ws" + strings.TrimPrefix(tg.server.URL, "http") + "/ws"
-}
-
-// generateToken creates a valid ES256 JWT for the given subject.
-func (tg *testGateway) generateToken(subject string, ttl time.Duration) string {
-	now := time.Now()
-	claims := jwt.MapClaims{
-		"iss":    "hotplex-worker",
-		"sub":    subject,
-		"aud":    "gateway",
-		"exp":    now.Add(ttl).Unix(),
-		"iat":    now.Unix(),
-		"nbf":    now.Unix(),
-		"jti":    uuid.NewString(),
-		"scopes": []string{"session:write"},
-	}
-	token := jwt.NewWithClaims(jwt.SigningMethodES256, claims)
-	s, err := token.SignedString(tg.jwtKey)
-	if err != nil {
-		panic(err)
-	}
-	return s
-}
-
-// ─── Helper: connect client ─────────────────────────────────────────────────
-
-func connectClient(t *testing.T, tg *testGateway, workerType string) *client.Client {
-	t.Helper()
-	token := tg.generateToken("test-user", 5*time.Minute)
-	c, err := client.New(context.Background(),
-		client.URL(tg.wsURL()),
-		client.WorkerType(workerType),
-		client.AuthToken(token),
-		client.APIKey("test-key"),
-	)
-	require.NoError(t, err)
-	return c
-}
-
-// collectEvents reads all events until done or timeout, returning them.
-func collectEvents(t *testing.T, ch <-chan client.Event, timeout time.Duration) []client.Event {
-	t.Helper()
-	var evts []client.Event
-	deadline := time.After(timeout)
-	for {
-		select {
-		case evt, ok := <-ch:
-			if !ok {
-				return evts
-			}
-			evts = append(evts, evt)
-			if evt.Type == client.EventDone || evt.Type == client.EventError {
-				return evts
-			}
-		case <-deadline:
-			return evts
-		}
-	}
-}
-
-// ─── Worker types for table-driven tests ────────────────────────────────────
-
-var allWorkerTypes = []struct {
-	name       string
-	workerType string
-}{
-	{"claude_code", string(worker.TypeClaudeCode)},
-	{"opencode_cli", string(worker.TypeOpenCodeCLI)},
-	{"opencode_server", string(worker.TypeOpenCodeSrv)},
-	{"acpx", string(worker.TypeACPX)},
-	{"pi_mono", string(worker.TypePimon)},
-}
-
-// ─── E2E Tests ──────────────────────────────────────────────────────────────
-
-func TestE2E_ConnectAndInit(t *testing.T) {
-	for _, wt := range allWorkerTypes {
-		t.Run(wt.name, func(t *testing.T) {
-			tg := setupTestGateway(t)
-			c := connectClient(t, tg, wt.workerType)
-
-			ack, err := c.Connect(context.Background())
-			require.NoError(t, err)
-			require.NotEmpty(t, ack.SessionID)
-			require.True(t, ack.State == client.StateCreated || ack.State == client.StateRunning,
-				"unexpected state: %s", ack.State)
-			require.Equal(t, events.Version, ack.ServerCaps.ProtocolVersion)
-			require.Equal(t, wt.workerType, ack.ServerCaps.WorkerType)
-
-			require.NoError(t, c.Close())
-		})
-	}
-}
-
-func TestE2E_SendInputReceiveEvents(t *testing.T) {
-	for _, wt := range allWorkerTypes {
-		t.Run(wt.name, func(t *testing.T) {
-			tg := setupTestGateway(t)
-			c := connectClient(t, tg, wt.workerType)
-
-			ack, err := c.Connect(context.Background())
-			require.NoError(t, err)
-			require.NotEmpty(t, ack.SessionID)
-
-			err = c.SendInput(context.Background(), "Hello, worker!")
-			require.NoError(t, err)
-
-			evts := collectEvents(t, c.Events(), 10*time.Second)
-
-			// Verify we got at least state, message.start, message.delta, message.end, done.
-			var hasState, hasMsgStart, hasDelta, hasMsgEnd, hasDone bool
-			for _, evt := range evts {
-				switch evt.Type {
-				case client.EventState:
-					hasState = true
-				case client.EventMessageStart:
-					hasMsgStart = true
-				case client.EventMessageDelta:
-					hasDelta = true
-				case client.EventMessageEnd:
-					hasMsgEnd = true
-				case client.EventDone:
-					hasDone = true
-				}
-			}
-			// State event may be lost due to a race between hub session
-			// registration and the worker's StateNotifier broadcast (especially
-			// under CI with slow scheduling). Log but don't fail.
-			if !hasState {
-				t.Log("state event not received (known race between hub registration and StateNotifier)")
-			}
-			require.True(t, hasMsgStart, "expected message.start event")
-			require.True(t, hasDelta, "expected message.delta event")
-			require.True(t, hasMsgEnd, "expected message.end event")
-			require.True(t, hasDone, "expected done event")
-
-			require.NoError(t, c.Close())
-		})
-	}
-}
-
-func TestE2E_PingPong(t *testing.T) {
-	tg := setupTestGateway(t)
-	c := connectClient(t, tg, string(worker.TypeClaudeCode))
-
-	_, err := c.Connect(context.Background())
-	require.NoError(t, err)
-
-	// Send a ping via the raw send method.
-	err = c.SendInput(context.Background(), "test") // trigger some activity
-	require.NoError(t, err)
-
-	// Drain events from input.
-	_ = collectEvents(t, c.Events(), 5*time.Second)
-
-	require.NoError(t, c.Close())
-}
-
-func TestE2E_SessionTerminate(t *testing.T) {
-	for _, wt := range allWorkerTypes {
-		t.Run(wt.name, func(t *testing.T) {
-			tg := setupTestGateway(t)
-			c := connectClient(t, tg, wt.workerType)
-
-			ack, err := c.Connect(context.Background())
-			require.NoError(t, err)
-			require.NotEmpty(t, ack.SessionID)
-
-			err = c.SendControl(context.Background(), client.ControlActionTerminate)
-			require.NoError(t, err)
-
-			// recvPump exits on error (IsTerminalEvent), so we check for error event.
-			evts := collectEvents(t, c.Events(), 5*time.Second)
-			var hasError bool
-			for _, evt := range evts {
-				if evt.Type == client.EventError {
-					hasError = true
-				}
-			}
-			require.True(t, hasError, "expected error event after terminate")
-
-			require.NoError(t, c.Close())
-		})
-	}
-}
-
-func TestE2E_SessionDelete(t *testing.T) {
-	for _, wt := range allWorkerTypes {
-		t.Run(wt.name, func(t *testing.T) {
-			tg := setupTestGateway(t)
-			c := connectClient(t, tg, wt.workerType)
-
-			ack, err := c.Connect(context.Background())
-			require.NoError(t, err)
-			sessionID := ack.SessionID
-
-			err = c.SendControl(context.Background(), client.ControlActionDelete)
-			require.NoError(t, err)
-
-			// Delete is async — poll until the session is removed.
-			require.Eventually(t, func() bool {
-				_, err := tg.sm.Get(sessionID)
-				return err != nil
-			}, 2*time.Second, 50*time.Millisecond, "session should be deleted")
-
-			require.NoError(t, c.Close())
-		})
-	}
-}
-
-// TestE2E_SessionReset verifies the reset control flow. After Connect the session
-// is RUNNING (Bridge.StartSession transitions CREATED→RUNNING). Reset attempts
-// RUNNING→RUNNING which is not a valid state transition, so the gateway responds
-// with an error event.
-func TestE2E_SessionReset(t *testing.T) {
-	for _, wt := range allWorkerTypes {
-		t.Run(wt.name, func(t *testing.T) {
-			tg := setupTestGateway(t)
-			c := connectClient(t, tg, wt.workerType)
-
-			ack, err := c.Connect(context.Background())
-			require.NoError(t, err)
-			require.NotEmpty(t, ack.SessionID)
-
-			// After Connect, session is RUNNING (via Bridge.StartSession).
-			// Reset attempts RUNNING→RUNNING which is invalid.
-			err = c.SendReset(context.Background(), "test reset")
-			require.NoError(t, err)
-
-			// Gateway sends error event for invalid transition.
-			evts := collectEvents(t, c.Events(), 5*time.Second)
-			var hasError bool
-			for _, evt := range evts {
-				if evt.Type == client.EventError {
-					hasError = true
-				}
-			}
-			require.True(t, hasError, "expected error event for reset from RUNNING state")
-
-			require.NoError(t, c.Close())
-		})
-	}
-}
-
-func TestE2E_SessionGC(t *testing.T) {
-	for _, wt := range allWorkerTypes {
-		t.Run(wt.name, func(t *testing.T) {
-			tg := setupTestGateway(t)
-			c := connectClient(t, tg, wt.workerType)
-
-			ack, err := c.Connect(context.Background())
-			require.NoError(t, err)
-			require.NotEmpty(t, ack.SessionID)
-
-			// Send GC directly (session is RUNNING from init).
-			err = c.SendGC(context.Background(), "test gc")
-			require.NoError(t, err)
-
-			// Verify session transitions to TERMINATED via session manager.
-			require.Eventually(t, func() bool {
-				si, err := tg.sm.Get(ack.SessionID)
-				if err != nil {
-					return false
-				}
-				return si.State == client.StateTerminated
-			}, 2*time.Second, 50*time.Millisecond, "session should be TERMINATED after GC")
-
-			require.NoError(t, c.Close())
-		})
-	}
-}
-
-func TestE2E_ResumeSession(t *testing.T) {
-	for _, wt := range allWorkerTypes {
-		t.Run(wt.name, func(t *testing.T) {
-			tg := setupTestGateway(t)
-
-			token := tg.generateToken("test-user", 5*time.Minute)
-
-			// First connection: create session.
-			c1, err := client.New(context.Background(),
-				client.URL(tg.wsURL()),
-				client.WorkerType(wt.workerType),
-				client.AuthToken(token),
-				client.APIKey("test-key"),
-			)
-			require.NoError(t, err)
-
-			ack1, err := c1.Connect(context.Background())
-			require.NoError(t, err)
-			sessionID := ack1.SessionID
-			require.NotEmpty(t, sessionID)
-
-			// Send some input.
-			err = c1.SendInput(context.Background(), "first message")
-			require.NoError(t, err)
-			_ = collectEvents(t, c1.Events(), 5*time.Second)
-
-			// Close first connection (session goes to IDLE).
-			require.NoError(t, c1.Close())
-
-			// Wait for session to transition to IDLE.
-			time.Sleep(200 * time.Millisecond)
-
-			// Second connection: resume with same session ID.
-			c2, err := client.New(context.Background(),
-				client.URL(tg.wsURL()),
-				client.WorkerType(wt.workerType),
-				client.AuthToken(token),
-				client.APIKey("test-key"),
-				client.ClientSessionID(sessionID),
-			)
-			require.NoError(t, err)
-
-			ack2, err := c2.Connect(context.Background())
-			require.NoError(t, err)
-			// Session ID may be derived differently via DeriveSessionKey,
-			// but the connection should succeed.
-			require.NotEmpty(t, ack2.SessionID)
-			require.True(t, ack2.State == client.StateRunning || ack2.State == client.StateIdle || ack2.State == client.StateCreated,
-				"unexpected resume state: %s", ack2.State)
-
-			require.NoError(t, c2.Close())
-		})
-	}
-}
-
-func TestE2E_MultipleWorkers(t *testing.T) {
-	tg := setupTestGateway(t)
-
-	var wg sync.WaitGroup
-	results := make(chan []client.Event, len(allWorkerTypes))
-
-	for _, wt := range allWorkerTypes {
-		wg.Add(1)
-		go func(workerType string) {
-			defer wg.Done()
-
-			c := connectClient(t, tg, workerType)
-			defer c.Close()
-
-			_, err := c.Connect(context.Background())
-			if err != nil {
-				results <- nil
-				return
-			}
-
-			err = c.SendInput(context.Background(), "hello from "+workerType)
-			if err != nil {
-				results <- nil
-				return
-			}
-
-			evts := collectEvents(t, c.Events(), 10*time.Second)
-			results <- evts
-		}(wt.workerType)
-	}
-
-	wg.Wait()
-	close(results)
-
-	// Verify each worker type received events.
-	count := 0
-	for evts := range results {
-		if evts != nil {
-			count++
-			var hasDone bool
-			for _, evt := range evts {
-				if evt.Type == client.EventDone {
-					hasDone = true
-				}
-			}
-			require.True(t, hasDone, "expected done event")
-		}
-	}
-	require.Equal(t, len(allWorkerTypes), count, "all workers should produce events")
-}
-
-func TestE2E_CloseGracefully(t *testing.T) {
-	tg := setupTestGateway(t)
-	c := connectClient(t, tg, string(worker.TypeClaudeCode))
-
-	_, err := c.Connect(context.Background())
-	require.NoError(t, err)
-
-	// Close should not block or panic.
-	done := make(chan struct{})
-	go func() {
-		require.NoError(t, c.Close())
-		close(done)
-	}()
-
-	select {
-	case <-done:
-		// Success.
-	case <-time.After(5 * time.Second):
-		t.Fatal("Close() blocked for too long")
-	}
-}
-
-func TestE2E_EventSeqMonotonic(t *testing.T) {
-	tg := setupTestGateway(t)
-	c := connectClient(t, tg, string(worker.TypeClaudeCode))
-
-	_, err := c.Connect(context.Background())
-	require.NoError(t, err)
-
-	err = c.SendInput(context.Background(), "check seq ordering")
-	require.NoError(t, err)
-
-	evts := collectEvents(t, c.Events(), 5*time.Second)
-
-	// Verify seq is monotonically increasing.
-	var lastSeq int64
-	for _, evt := range evts {
-		if evt.Seq > 0 && evt.Type != client.EventInitAck {
-			require.Greater(t, evt.Seq, lastSeq,
-				"seq should be monotonic: last=%d current=%d type=%s", lastSeq, evt.Seq, evt.Type)
-			lastSeq = evt.Seq
-		}
-	}
-
-	require.NoError(t, c.Close())
-}
-
-// TestE2E_MultipleInputsSequential verifies sequential input handling.
-// After the first input produces a done event, the client's recvPump exits
-// (IsTerminalEvent). Subsequent inputs are accepted by the gateway but events
-// cannot be received on the same connection.
-func TestE2E_MultipleInputsSequential(t *testing.T) {
-	tg := setupTestGateway(t)
-	c := connectClient(t, tg, string(worker.TypeClaudeCode))
-
-	_, err := c.Connect(context.Background())
-	require.NoError(t, err)
-
-	// First input: produces full event sequence including done.
-	err = c.SendInput(context.Background(), "message 0")
-	require.NoError(t, err)
-	evts := collectEvents(t, c.Events(), 5*time.Second)
-	var hasDone bool
-	for _, evt := range evts {
-		if evt.Type == client.EventDone {
-			hasDone = true
-		}
-	}
-	require.True(t, hasDone, "expected done event for first input")
-
-	// Second input: accepted by gateway (session stays RUNNING),
-	// but client's recvPump has exited on the first done event.
-	err = c.SendInput(context.Background(), "message 1")
-	require.NoError(t, err)
-
-	require.NoError(t, c.Close())
-}
-
-func TestE2E_DoneDataSuccess(t *testing.T) {
-	tg := setupTestGateway(t)
-	c := connectClient(t, tg, string(worker.TypeClaudeCode))
-
-	_, err := c.Connect(context.Background())
-	require.NoError(t, err)
-
-	err = c.SendInput(context.Background(), "check done data")
-	require.NoError(t, err)
-
-	evts := collectEvents(t, c.Events(), 5*time.Second)
-
-	var doneEvt *client.Event
-	for i := range evts {
-		if evts[i].Type == client.EventDone {
-			doneEvt = &evts[i]
-			break
-		}
-	}
-	require.NotNil(t, doneEvt, "expected done event")
-
-	// Verify done data structure.
-	data, ok := doneEvt.Data.(map[string]any)
-	require.True(t, ok)
-	success, _ := data["success"].(bool)
-	require.True(t, success, "done.success should be true")
-
-	require.NoError(t, c.Close())
-}
-
-func TestE2E_MessageDeltaContent(t *testing.T) {
-	tg := setupTestGateway(t)
-	c := connectClient(t, tg, string(worker.TypeClaudeCode))
-
-	_, err := c.Connect(context.Background())
-	require.NoError(t, err)
-
-	err = c.SendInput(context.Background(), "Hello World")
-	require.NoError(t, err)
-
-	evts := collectEvents(t, c.Events(), 5*time.Second)
-
-	// Collect all delta content.
-	var deltaContent string
-	for _, evt := range evts {
-		if evt.Type == client.EventMessageDelta {
-			if data, ok := evt.Data.(map[string]any); ok {
-				if content, ok := data["content"].(string); ok {
-					deltaContent += content
-				}
-			}
-		}
-	}
-	require.Contains(t, deltaContent, "Hello World",
-		"delta content should contain the input text")
-
-	require.NoError(t, c.Close())
-}
-
-// TestE2E_AuthFailure verifies that connecting without credentials fails.
-func TestE2E_AuthFailure(t *testing.T) {
-	tg := setupTestGateway(t)
-
-	// Create client without API key — should fail at HTTP upgrade.
-	c, err := client.New(context.Background(),
-		client.URL(tg.wsURL()),
-		client.WorkerType(string(worker.TypeClaudeCode)),
-	)
-	require.NoError(t, err)
-
-	_, err = c.Connect(context.Background())
-	require.Error(t, err, "expected auth failure without API key")
-}
-
-// TestE2E_LargeInput verifies that the gateway handles large messages.
-func TestE2E_LargeInput(t *testing.T) {
-	tg := setupTestGateway(t)
-	c := connectClient(t, tg, string(worker.TypeClaudeCode))
-
-	_, err := c.Connect(context.Background())
-	require.NoError(t, err)
-
-	largeContent := strings.Repeat("x", 10000)
-	err = c.SendInput(context.Background(), largeContent)
-	require.NoError(t, err)
-
-	evts := collectEvents(t, c.Events(), 5*time.Second)
-	var hasDone bool
-	for _, evt := range evts {
-		if evt.Type == client.EventDone {
-			hasDone = true
-		}
-	}
-	require.True(t, hasDone, "expected done event for large input")
-
-	require.NoError(t, c.Close())
-}
-
-// ─── Helper: parse event data ───────────────────────────────────────────────
